deleteing a story will trigger a confirmation window, other window alters for errors, creating a new character already displayed on jumbotron will add edit buttons

+++ src/story.js
@@ -100,7 +100,9 @@ class Story{
         deleteButton.innerText = "Delete"
         buttons.append(deleteButton)
         deleteButton.addEventListener('click', () => {
-            this.deleteStory();
+            if (window.confirm("Are you sure you want to delete this story and all related characters?")){
+                this.deleteStory();
+            }
         })
     }
